게시글 검증 처리 고도화

+++ src/main/java/com/wanted/preonboarding/service/impl/BoardServiceImpl.java
@@ -5,6 +5,8 @@ import com.wanted.preonboarding.entity.board.Board;
 import com.wanted.preonboarding.entity.board.dto.BoardInfoDto;
 import com.wanted.preonboarding.entity.board.dto.BoardRequestDto;
 import com.wanted.preonboarding.entity.user.User;
+import com.wanted.preonboarding.exception.board.imple.BoardAuthorNotMatchException;
+import com.wanted.preonboarding.exception.board.imple.BoardNotFoundException;
 import com.wanted.preonboarding.exception.user.impl.UserNotFoundException;
 import com.wanted.preonboarding.repository.BoardRepository;
 import com.wanted.preonboarding.repository.UserRepository;
@@ -40,12 +42,8 @@ public class BoardServiceImpl implements BoardService {
 
     @Override
     public void updateBoard(String token, Long boardId, BoardRequestDto boardRequestDto) {
-        Board board = boardRepository.findById(boardId).orElseThrow(() -> new RuntimeException("게시글이 존재하지 않습니다."));
-        User user = userRepository.getByEmail(jwtTokenProvider.getUsername(token)).orElseThrow(UserNotFoundException::new);
-
-        if (!board.getUser().getEmail().equals(user.getEmail())) {
-            throw new RuntimeException("게시글 작성자가 아닙니다.");
-        }
+        Board board = boardRepository.findById(boardId).orElseThrow(BoardNotFoundException::new);
+        validateAuthor(token, board);
 
         board.update(boardRequestDto);
         boardRepository.save(board);
@@ -53,17 +51,26 @@ public class BoardServiceImpl implements BoardService {
 
     @Override
     public void deleteBoard(String token, Long boardId) {
-        Board board = boardRepository.findById(boardId).orElseThrow(() -> new RuntimeException("게시글이 존재하지 않습니다."));
+        Board board = boardRepository.findById(boardId).orElseThrow(BoardNotFoundException::new);
+        validateAuthor(token, board);
         boardRepository.delete(board);
     }
 
     @Override
     public BoardInfoDto getBoard(Long boardId) {
-        return boardRepository.findById(boardId).orElseThrow(() -> new RuntimeException("게시글이 존재하지 않습니다.")).toBoardInfoDto();
+        return boardRepository.findById(boardId).orElseThrow(BoardNotFoundException::new).toBoardInfoDto();
     }
 
     @Override
     public Page<BoardInfoDto> getPosts(Pageable pageable) {
         return boardRepository.findAll(pageable).map(Board::toBoardInfoDto);
+    }
+
+    private void validateAuthor(String token, Board board) {
+        String username = jwtTokenProvider.getUsername(token);
+        User author = userRepository.getByEmail(username).orElseThrow(UserNotFoundException::new);
+        if (!board.getUser().getEmail().equals(author.getEmail())) {
+            throw new BoardAuthorNotMatchException();
+        }
     }
 }
